move NewContainer to ctr/containers
